Revert "feat(performance): Lazy load grids on collect/id, collection/id, artist/id"
