+++ src/Apps/Artist/Components/ArtistHeader/ArtistHeaderImage.tsx
@@ -53,13 +53,11 @@ export const ArtistHeaderImage: FC<ArtistHeaderImageProps> = ({
           {...rest}
         >
           <Image
-            key={desktop.src}
             src={desktop.src}
             srcSet={desktop.srcSet}
             width="100%"
             height="100%"
             style={{ objectFit: "cover" }}
-            lazyLoad
           />
         </ResponsiveBox>
       </Media>
