+++ src/Apps/Artist/Routes/WorksForSale/Components/ArtistArtworkFilter.tsx
@@ -93,7 +93,10 @@ export const ArtistArtworkFilterRefetchContainer = createRefetchContainer(
   {
     artist: graphql`
       fragment ArtistArtworkFilter_artist on Artist
-        @argumentDefinitions(input: { type: "FilterArtworksInput" }) {
+        @argumentDefinitions(
+          includeBlurHash: { type: "Boolean!", defaultValue: true }
+          input: { type: "FilterArtworksInput" }
+        ) {
         name
         counts {
           partner_shows: partnerShows
@@ -109,6 +112,7 @@ export const ArtistArtworkFilterRefetchContainer = createRefetchContainer(
             total(format: "0,0")
           }
           ...ArtworkFilterArtworkGrid_filtered_artworks
+            @arguments(includeBlurHash: $includeBlurHash)
         }
         internalID
         name
@@ -120,9 +124,11 @@ export const ArtistArtworkFilterRefetchContainer = createRefetchContainer(
     query ArtistArtworkFilterQuery(
       $artistID: String!
       $input: FilterArtworksInput
+      $includeBlurHash: Boolean!
     ) {
       artist(id: $artistID) {
-        ...ArtistArtworkFilter_artist @arguments(input: $input)
+        ...ArtistArtworkFilter_artist
+          @arguments(input: $input, includeBlurHash: $includeBlurHash)
       }
     }
   `
