+++ src/Apps/Collect/Routes/Collect/index.tsx
@@ -1,5 +1,4 @@
 import { Box, Separator, Spacer, Text, Flex } from "@artsy/palette"
-import StaticContainer from "found/StaticContainer"
 import { Match, Router } from "found"
 import * as React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
@@ -10,7 +9,6 @@ import { BreadCrumbList } from "Components/Seo/BreadCrumbList"
 import { getMetadata, Medium, Color } from "./Utils/getMetadata"
 import { Collect_marketingCollections$data } from "__generated__/Collect_marketingCollections.graphql"
 import { collectRoutes_ArtworkFilterQuery$data } from "__generated__/collectRoutes_ArtworkFilterQuery.graphql"
-import { CollectArtworkFilterQuery } from "__generated__/CollectArtworkFilterQuery.graphql"
 import { CollectionsHubsNavFragmentContainer as CollectionsHubsNav } from "Components/CollectionsHubsNav"
 import { ArtworkFilter } from "Components/ArtworkFilter"
 import { RouterLink } from "System/Components/RouterLink"
@@ -21,14 +19,12 @@ import {
 import { useSystemContext } from "System/Hooks/useSystemContext"
 import { useRouter } from "System/Hooks/useRouter"
 import { MetaTags } from "Components/MetaTags"
-import { initializeVariablesWithFilterState } from "Apps/Collect/collectRoutes"
-import { SystemQueryRenderer } from "System/Relay/SystemQueryRenderer"
-import { ArtworkFilterPlaceholder } from "Components/ArtworkFilter/ArtworkFilterPlaceholder"
 
 export interface CollectAppProps {
   match: Match
   router: Router
   marketingCollections: Collect_marketingCollections$data
+  viewer: collectRoutes_ArtworkFilterQuery$data["viewer"]
   filterArtworks: collectRoutes_ArtworkFilterQuery$data["filterArtworks"]
 }
 
@@ -36,10 +32,10 @@ export const CollectApp: React.FC<CollectAppProps> = ({
   filterArtworks,
   marketingCollections,
   match: { location, params },
+  viewer,
 }) => {
-  const { silentReplace, match } = useRouter()
+  const { silentReplace } = useRouter()
   const { userPreferences } = useSystemContext()
-
   const medium = params?.medium as Medium
   const color = params?.color as Color
   const { description, breadcrumbTitle, title } = getMetadata({
@@ -101,78 +97,28 @@ export const CollectApp: React.FC<CollectAppProps> = ({
         </Box>
 
         <Box>
-          {/* TODO: Figure out why rerenders trigger refetches here, requiring
-              the static container to freeze rendering during route transitions. */}
-          <StaticContainer shouldUpdate={!!match.elements}>
-            <SystemQueryRenderer<CollectArtworkFilterQuery>
-              query={graphql`
-                query CollectArtworkFilterQuery(
-                  $input: FilterArtworksInput
-                  $aggregations: [ArtworkAggregation]
-                  $shouldFetchCounts: Boolean!
-                ) {
-                  viewer {
-                    ...ArtworkFilter_viewer @arguments(input: $input)
-                    artworksConnection(
-                      aggregations: $aggregations
-                      input: $input
-                    ) {
-                      counts @include(if: $shouldFetchCounts) {
-                        followedArtists
-                      }
-                      aggregations {
-                        slice
-                        counts {
-                          value
-                          name
-                          count
-                        }
-                      }
-                    }
-                  }
-                }
-              `}
-              variables={initializeVariablesWithFilterState(
-                match.params,
-                match
-              )}
-              fetchPolicy="store-and-network"
-              placeholder={<ArtworkFilterPlaceholder />}
-              render={({ props }) => {
-                if (!props?.viewer) {
-                  return <ArtworkFilterPlaceholder />
-                }
+          <ArtworkFilter
+            viewer={viewer}
+            aggregations={
+              viewer?.artworksConnection
+                ?.aggregations as SharedArtworkFilterContextProps["aggregations"]
+            }
+            counts={viewer?.artworksConnection?.counts as Counts}
+            filters={location.query as any}
+            sortOptions={[
+              { text: "Recommended", value: "-decayed_merch" },
+              { text: "Recently Updated", value: "-partner_updated_at" },
+              { text: "Recently Added", value: "-published_at" },
+              { text: "Artwork Year (Descending)", value: "-year" },
+              { text: "Artwork Year (Ascending)", value: "year" },
+            ]}
+            onChange={filters => {
+              const url = buildUrlForCollectApp(filters)
 
-                return (
-                  <ArtworkFilter
-                    viewer={props.viewer}
-                    aggregations={
-                      props.viewer?.artworksConnection
-                        ?.aggregations as SharedArtworkFilterContextProps["aggregations"]
-                    }
-                    counts={props.viewer?.artworksConnection?.counts as Counts}
-                    filters={location.query as any}
-                    sortOptions={[
-                      { text: "Recommended", value: "-decayed_merch" },
-                      {
-                        text: "Recently Updated",
-                        value: "-partner_updated_at",
-                      },
-                      { text: "Recently Added", value: "-published_at" },
-                      { text: "Artwork Year (Descending)", value: "-year" },
-                      { text: "Artwork Year (Ascending)", value: "year" },
-                    ]}
-                    onChange={filters => {
-                      const url = buildUrlForCollectApp(filters)
-
-                      silentReplace(url)
-                    }}
-                    userPreferredMetric={userPreferences?.metric}
-                  />
-                )
-              }}
-            />
-          </StaticContainer>
+              silentReplace(url)
+            }}
+            userPreferredMetric={userPreferences?.metric}
+          />
         </Box>
       </FrameWithRecentlyViewed>
     </>
