+++ src/Apps/Collect/Routes/Collection/Components/Header/index.tsx
@@ -1,3 +1,4 @@
+import { ContextModule } from "@artsy/cohesion"
 import {
   Text,
   GridColumns,
@@ -11,10 +12,12 @@ import { Header_artworks$data } from "__generated__/Header_artworks.graphql"
 import { Header_collection$data } from "__generated__/Header_collection.graphql"
 import { CollectionDefaultHeaderFragmentContainer } from "Apps/Collect/Routes/Collection/Components/Header/DefaultHeader"
 import { Link } from "found"
+import { compact, filter, take } from "lodash"
 import * as React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { FullBleedHeader } from "Components/FullBleedHeader/FullBleedHeader"
 import { slugify } from "underscore.string"
+import { EntityHeaderArtistFragmentContainer } from "Components/EntityHeaders/EntityHeaderArtist"
 
 export interface CollectionHeaderProps {
   collection: Header_collection$data
@@ -25,6 +28,15 @@ export const CollectionHeader: React.FC<CollectionHeaderProps> = ({
   artworks,
   collection,
 }) => {
+  const merchandisableArtists = artworks?.merchandisableArtists ?? []
+  const hasMultipleArtists = merchandisableArtists.length > 1
+
+  const featuredArtists = getFeaturedArtists(
+    12,
+    collection,
+    merchandisableArtists
+  )
+
   return (
     <>
       {collection.headerImage && (
@@ -74,11 +86,41 @@ export const CollectionHeader: React.FC<CollectionHeaderProps> = ({
       )}
 
       <Spacer y={6} />
+
+      {collection.showFeaturedArtists && featuredArtists && hasMultipleArtists && (
+        <>
+          <Text variant="lg-display" mb={4}>
+            Featured Artists
+          </Text>
+
+          <GridColumns>
+            {featuredArtists.map(artist => {
+              if (!artist.name) return
+
+              return (
+                <Column
+                  span={[12, 6, 3, 3]}
+                  key={artist.internalID}
+                  data-test={ContextModule.featuredArtistsRail}
+                  display="flex"
+                  alignItems="center"
+                >
+                  <EntityHeaderArtistFragmentContainer
+                    artist={artist}
+                    width="100%"
+                    alignItems="flex-start"
+                  />
+                </Column>
+              )
+            })}
+          </GridColumns>
+        </>
+      )}
     </>
   )
 }
 
-export const CollectionHeaderFragmentContainer = createFragmentContainer(
+export const CollectionFilterFragmentContainer = createFragmentContainer(
   CollectionHeader,
   {
     collection: graphql`
@@ -110,3 +152,34 @@ export const CollectionHeaderFragmentContainer = createFragmentContainer(
     `,
   }
 )
+
+export const getFeaturedArtists = (
+  artistsCount: number,
+  collection: Header_collection$data,
+  merchandisableArtists: Header_artworks$data["merchandisableArtists"]
+) => {
+  if ((collection?.query?.artistIDs?.length ?? 0) > 0) {
+    return compact(
+      filter(merchandisableArtists, artist =>
+        (collection?.query?.artistIDs ?? []).includes(
+          artist?.internalID as string
+        )
+      )
+    )
+  }
+
+  if ((merchandisableArtists?.length ?? 0) > 0) {
+    const filteredArtistsIds = compact(
+      merchandisableArtists?.filter(artist => {
+        return !collection?.featuredArtistExclusionIds?.includes(
+          artist?.internalID as string
+        )
+      })
+    )
+
+    return take(filteredArtistsIds, artistsCount)
+  }
+
+  // No artists
+  return []
+}
