+++ src/Apps/Collect/Routes/Collection/index.tsx
@@ -1,8 +1,6 @@
 import { Spacer } from "@artsy/palette"
-import StaticContainer from "found/StaticContainer"
 import { Collection_collection$data } from "__generated__/Collection_collection.graphql"
-import { CollectionArtworksQuery } from "__generated__/CollectionArtworksQuery.graphql"
-import { CollectionHeaderFragmentContainer } from "Apps/Collect/Routes/Collection/Components/Header"
+import { CollectionFilterFragmentContainer as CollectionHeader } from "Apps/Collect/Routes/Collection/Components/Header"
 import { FrameWithRecentlyViewed } from "Components/FrameWithRecentlyViewed"
 import { RelatedCollectionsRailQueryRenderer } from "Components/RelatedCollectionsRail/RelatedCollectionsRail"
 import { BreadCrumbList } from "Components/Seo/BreadCrumbList"
@@ -24,11 +22,6 @@ import {
   withSystemContext,
 } from "System/Contexts/SystemContext"
 import { ArtworkGridContextProvider } from "Components/ArtworkGrid/ArtworkGridContext"
-import { SystemQueryRenderer } from "System/Relay/SystemQueryRenderer"
-import { initializeVariablesWithFilterState } from "Apps/Collect/collectRoutes"
-import { useRouter } from "System/Hooks/useRouter"
-import { ArtworkFilterPlaceholder } from "Components/ArtworkFilter/ArtworkFilterPlaceholder"
-import { CollectionFeaturedArtistsFragmentContainer } from "Apps/Collect/Routes/Collection/Components/Header/CollectionFeaturedArtists"
 
 interface CollectionAppProps extends SystemContextProps {
   collection: Collection_collection$data
@@ -39,8 +32,6 @@ interface CollectionAppProps extends SystemContextProps {
 export const CollectionApp: React.FC<CollectionAppProps> = props => {
   const { collection } = props
 
-  const { match } = useRouter()
-
   if (!collection) return <ErrorPage code={404} />
 
   const {
@@ -49,6 +40,7 @@ export const CollectionApp: React.FC<CollectionAppProps> = props => {
     headerImage,
     descriptionMarkdown,
     fallbackHeaderImage,
+    artworksConnection,
   } = collection
 
   const metadataDescription = descriptionMarkdown
@@ -89,111 +81,41 @@ export const CollectionApp: React.FC<CollectionAppProps> = props => {
         ]}
       />
 
-      {collection.artworksConnection && (
-        <CollectionHeaderFragmentContainer
-          collection={collection}
-          artworks={collection.artworksConnection}
-        />
-      )}
+      {/* @ts-expect-error PLEASE_FIX_ME_STRICT_NULL_CHECK_MIGRATION */}
+      <CollectionHeader collection={collection} artworks={artworksConnection} />
 
-      <ArtworkGridContextProvider hideSignals={hideSignals}>
-        {/* TODO: Figure out why rerenders trigger refetches here, requiring
-              the static container to freeze rendering during route transitions. */}
-        <StaticContainer shouldUpdate={!!match.elements}>
-          <SystemQueryRenderer<CollectionArtworksQuery>
-            query={graphql`
-              query CollectionArtworksQuery(
-                $slug: String!
-                $aggregations: [ArtworkAggregation]
-                $input: FilterArtworksInput!
-                $shouldFetchCounts: Boolean!
-              ) {
-                marketingCollection(slug: $slug) {
-                  ...CollectionFeaturedArtists_collection
-                  ...CollectionArtworksFilter_collection
-                    @arguments(input: $input)
+      <FrameWithRecentlyViewed>
+        {showCollectionHubs && (
+          <>
+            <Spacer y={6} />
 
-                  artworksConnection(
-                    aggregations: $aggregations
-                    includeMediumFilterInAggregation: true
-                    first: 20
-                    sort: "-decayed_merch"
-                  ) {
-                    ...CollectionFeaturedArtists_artworks
-                    counts @include(if: $shouldFetchCounts) {
-                      followedArtists
-                    }
-                    aggregations {
-                      slice
-                      counts {
-                        value
-                        name
-                        count
-                      }
-                    }
-                  }
-                }
-              }
-            `}
-            variables={initializeVariablesWithFilterState(match.params, match)}
-            placeholder={<ArtworkFilterPlaceholder pt={6} />}
-            render={({ error, props }) => {
-              if (error) {
-                console.error("[collection]: Error loading artwork grid", error)
-                return null
-              }
+            <CollectionsHubRails
+              linkedCollections={collection.linkedCollections}
+            />
+          </>
+        )}
 
-              if (!props || !props.marketingCollection) {
-                return <ArtworkFilterPlaceholder pt={6} />
-              }
+        <Spacer y={6} />
 
-              return (
-                <>
-                  {props.marketingCollection.artworksConnection && (
-                    <CollectionFeaturedArtistsFragmentContainer
-                      collection={props.marketingCollection}
-                      artworks={props.marketingCollection.artworksConnection}
-                    />
-                  )}
-
-                  <FrameWithRecentlyViewed>
-                    {showCollectionHubs && (
-                      <>
-                        <Spacer y={6} />
-
-                        <CollectionsHubRails
-                          linkedCollections={collection.linkedCollections}
-                        />
-                      </>
-                    )}
-
-                    <Spacer y={6} />
-
-                    <CollectionArtworksFilterRefetchContainer
-                      collection={props.marketingCollection}
-                      aggregations={
-                        props.marketingCollection?.artworksConnection
-                          ?.aggregations as SharedArtworkFilterContextProps["aggregations"]
-                      }
-                      counts={
-                        props.marketingCollection.artworksConnection
-                          ?.counts as Counts
-                      }
-                    />
-                    {collection.linkedCollections.length === 0 && (
-                      <>
-                        <Spacer y={6} />
-
-                        <RelatedCollectionsRailQueryRenderer slug={slug} />
-                      </>
-                    )}
-                  </FrameWithRecentlyViewed>
-                </>
-              )
-            }}
+        <ArtworkGridContextProvider hideSignals={hideSignals}>
+          <CollectionArtworksFilterRefetchContainer
+            collection={collection}
+            aggregations={
+              collection.artworksConnection
+                ?.aggregations as SharedArtworkFilterContextProps["aggregations"]
+            }
+            counts={collection.artworksConnection?.counts as Counts}
           />
-        </StaticContainer>
-      </ArtworkGridContextProvider>
+        </ArtworkGridContextProvider>
+
+        {collection.linkedCollections.length === 0 && (
+          <>
+            <Spacer y={6} />
+
+            <RelatedCollectionsRailQueryRenderer slug={slug} />
+          </>
+        )}
+      </FrameWithRecentlyViewed>
     </>
   )
 }
@@ -214,7 +136,11 @@ export const CollectionFragmentContainer = createFragmentContainer(
   {
     collection: graphql`
       fragment Collection_collection on MarketingCollection
-        @argumentDefinitions(aggregations: { type: "[ArtworkAggregation]" }) {
+        @argumentDefinitions(
+          aggregations: { type: "[ArtworkAggregation]" }
+          input: { type: "FilterArtworksInput" }
+          shouldFetchCounts: { type: "Boolean!", defaultValue: false }
+        ) {
         ...Header_collection
         # TODO: Description should implement markdown which accepts a format argument
         descriptionMarkdown
@@ -244,11 +170,24 @@ export const CollectionFragmentContainer = createFragmentContainer(
         artworksConnection(
           aggregations: $aggregations
           includeMediumFilterInAggregation: true
-          first: 5
+          first: 20
           sort: "-decayed_merch"
         ) {
           ...Header_artworks
+          counts @include(if: $shouldFetchCounts) {
+            followedArtists
+          }
+          aggregations {
+            slice
+            counts {
+              value
+              name
+              count
+            }
+          }
         }
+
+        ...CollectionArtworksFilter_collection @arguments(input: $input)
       }
     `,
   }
